+++ theme/functions.php
@@ -17,6 +17,7 @@ require_once get_template_directory() . '/inc/assets.php';
 require_once get_template_directory() . '/inc/template-tags.php';
 require_once get_template_directory() . '/inc/patterns.php';
 require_once get_template_directory() . '/inc/portal.php';
+require_once get_template_directory() . '/inc/teacher-portal.php';
 
 /**
  * Randomise the Delnavazan homepage article Query Loop.